+++ src/Sulu/Component/Content/Document/Subscriber/ResourceSegmentSubscriber.php
@@ -104,8 +104,11 @@ class ResourceSegmentSubscriber implements EventSubscriberInterface
             return;
         }
 
-        $property = $this->getResourceSegmentProperty($document);
+        if ($document instanceof RedirectTypeBehavior && $document->getRedirectType() !== RedirectType::NONE) {
+            return;
+        }
 
+        $property = $this->getResourceSegmentProperty($document);
         $document->getStructure()->getProperty(
             $property->getName()
         )->setValue($document->getResourceSegment());