+++ CHANGELOG.md
@@ -73,6 +73,11 @@ CHANGELOG for Sulu
     * BUGFIX      #1365 [ContactBundle]   Fixed bug caused by new instance name of datagrid
     * BUGFIX      #1136 [MediaBundle]     Fixed image scale forceRatio parameter for none squared image formats
 
+* 1.0.11 (2015-09-23)
+    * HOTFIX      #1596 [GeneratorBundle] Fixed sulu bundle generator path generation
+    * HOTFIX      #1615 [Content]         Fixed resource segment subscriber for order internal link
+    * HOTFIX      #1612 [Content]         Fixed non copy content of snippet
+
 * 1.0.10 (2015-09-17)
     * HOTFIX      #1594 [Website] Fixed website request analyzer
     * HOTFIX      #1594 [Website] Fixed trailing slash for homepage